Move view_trial package to participants
